+++ Client/src/pages/issue/IssuePage.jsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import styled from 'styled-components';
 import { Link } from 'react-router-dom';
-import MenuButton from '../../components/MenuButton';
+import MenuButton from '../../components/button/MenuButton';
+import NewButton from '../../components/button/NewButton';
 
 const Container = styled.div`
   display: flex;
@@ -29,24 +30,6 @@ const IssueList = styled.div`
 const MenuDiv = styled.div`
   display: flex;
   flex: ${(props) => props.flex};
-`;
-
-const IssueButton = styled.button`
-  width: 120px;
-  box-shadow: 0px 1px 0px 0px #3dc21b;
-  background: linear-gradient(to bottom, #44c767 5%, #5cbf2a 100%);
-  background-color: #44c767;
-  border-radius: 6px;
-  border: 1px solid #18ab29;
-  display: inline-block;
-  cursor: pointer;
-  color: white;
-  font-size: 15px;
-  font-weight: bold;
-  padding: 10px;
-  text-decoration: none;
-  text-shadow: 0px -1px 0px #2f6627;
-  margin-left: 20px;
 `;
 
 const FilterButton = styled.button`
@@ -95,9 +78,7 @@ export default function IssuePage() {
           ></MenuButton>
         </MenuDiv>
         <MenuDiv flex="0">
-          <Link to="/issue/post">
-            <IssueButton>New issue</IssueButton>
-          </Link>
+          <NewButton link="/issue/post" name="issue" />
         </MenuDiv>
       </IssueMenu>
       <IssueList>
